apps/backend/pyproject.toml
[project]
name = "cba-platform-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "httpx>=0.28.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "redis>=5.2.1",
  "rq>=2.0.0",
  "structlog>=24.4.0",
  "sentry-sdk[fastapi]>=2.19.0",
  "opentelemetry-sdk>=1.29.0",
  "opentelemetry-instrumentation-fastapi>=0.50b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.50b0",
  "opentelemetry-exporter-otlp>=1.29.0",
  "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "pytest-cov>=6.0.0",
  "httpx>=0.28.0",
  "factory-boy>=3.3.1",
  "ruff>=0.8.0",
  "mypy>=1.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
